split up gateway.js file
