+++ lib/gateway/request.js
@@ -1,0 +1,64 @@
+'use strict'
+const http = require('http')
+const https = require('https')
+const pump = require('pump')
+
+function end (req, body, cb) {
+  if (!body || typeof body === 'string' || body instanceof Uint8Array) {
+    req.end(body)
+  } else if (body.pipe) {
+    pump(body, req, err => {
+      if (err) cb(err)
+    })
+  } else {
+    cb(new Error(`type unsupported for body: ${body.constructor}`))
+  }
+}
+
+function agentOption (opts) {
+  return {
+    keepAlive: true,
+    keepAliveMsecs: opts.keepAliveMsecs || 60 * 1000, // 1 minute
+    maxSockets: opts.maxSockets || 2048,
+    maxFreeSockets: opts.maxFreeSockets || 2048,
+    rejectUnauthorized: opts.rejectUnauthorized
+  }
+}
+
+module.exports = function buildRequest (opts) {
+  const agents = {
+    'http:': new http.Agent(agentOption(opts)),
+    'https:': new https.Agent(agentOption(opts))
+  }
+
+  const requests = {
+    'http:': http,
+    'https:': https
+  }
+
+  function close () {
+    agents['http:'].destroy()
+    agents['https:'].destroy()
+  }
+
+  function request (opts, done) {
+    const req = requests[opts.url.protocol].request({
+      method: opts.method,
+      port: opts.url.port,
+      path: opts.url.pathname + (opts.qs || ''),
+      hostname: opts.url.hostname,
+      headers: opts.headers,
+      agent: agents[opts.url.protocol]
+    })
+    req.on('error', done)
+    req.on('response', res => {
+      done(null, { statusCode: res.statusCode, headers: res.headers, stream: res })
+    })
+    end(req, opts.body, done)
+  }
+
+  return {
+    request,
+    close
+  }
+}
