+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "unit": "tap --no-esm --no-jsx --100 -J test/*.js",
-    "cov": "tap  --no-esm --no-jsx --coverage-report=html -J test/*.js",
+    "unit": "tap --no-esm --no-jsx --100 -J test/*.js test/gateway/*.js",
+    "cov": "tap  --no-esm --no-jsx --coverage-report=html -J test/*.js test/gateway/*.js",
     "lint": "npm run lint:standard && npm run lint:typescript",
     "lint:standard": "standard | snazzy",
     "lint:typescript": "standard --parser @typescript-eslint/parser --plugin @typescript-eslint/eslint-plugin test/types/*.ts",
